oauth err solved

+++ src/controller/page.js
@@ -7,6 +7,11 @@ const newPage = async (req, res) => {
         .status(400)
         .send({ error: "admin is not authorized to create page" });
     }
+    if (req.user.status === "blocked"){
+      return res
+      .status(400)
+      .send({ error: "user is blocked" })
+    }
     const page = new Page({ ...req.body, owner: req.user._id });
     const savedPage = await page.save();
     res.status(201).json({ message: "Page created successfully", savedPage });
@@ -53,6 +58,11 @@ const deletePage = async (req, res) => {
   const userId = req.user._id;
 
   try {
+    if (req.user.status === "blocked"){
+      return res
+      .status(400)
+      .send({ error: "user is blocked" })
+    }
     const deletedPage = await Page.findOneAndDelete({
       _id: pageId,
       owner: userId,
@@ -71,6 +81,11 @@ const deletePage = async (req, res) => {
 };
 const updatePage = async (req, res) => {
   try {
+    if (req.user.status === "blocked"){
+      return res
+      .status(400)
+      .send({ error: "user is blocked" })
+    }
     const pageId = req.params.id;
     const userId = req.user._id;
